kernel: ChangedStorage type cleanup
Changed storage as well as newStorage are based on ObjectStorage template
type. Let's hide ObjectStorage and typedef it to ChangedStorage resp.
RefStorage so that usage is easier and more flexible.

This change also fixes compile-time error on VS2008 because of missing const
for Ref in Iterator definition in saveChanges, cleanUp and getNumObjects (which
has been introduced lately by API cleanup changes).

Thanks to Jozo for pointing this out