CXref::stream proper cleanup
Fixes bt#329

CPdf::getInstance deallocates stream if the CPdf instance creation fails.
This is not 100% correct because CPdf can fail from multiple places.

One possibility is that CXRef constructor fails. In this case, we have to
deallocate stream, because there is notbody else who can do it (CXref
destructor is not called and this is the only place where we deallocate
stream).

Second option is that XRefWriter constructor fails. In that case CXref
destructor is called and so stream is deallocated.

CPdf::getInstance checks whether instance==NULL and deletes stream in such a
case. This has to be changed because we have no chance to figure out the
source of exception and moreover the cleanup should be done on the proper
place. In this case this is CXref constructor which has to deallocate stream
before it throws. This patch adds CXref::init method which does basically
everything what constructor is supposed to do and constructor can wrapp
this method by exception handler which deallocates the stream.


This way we will keep deallocation in one module. XRefWriter and higher layers
don't have to care about stream at all.

Throwing from constructors is almost always a bad idea, but if we want to
have CPdf::getInstance semantic (return valid and prepared pdf or nothing)
simple then this is the only solution.