xpdf: reset stream in constructor
We used to trust that the given stream to the constructor points to the
beginning of the document data. This is, however, not tru when XRef is
created in PDFDoc::setup() method which calls PDFDoc::checkHeader which
scans the first headerSearchSize bytes for the PDF header so the real
position in the file is headerSearchSize.

Moreover PDFDoc::checkHeader calls str->moveStart which moves the
current position pointer without changing the real position in the file.
This causes XRef constructor not being able to find PDF header thus it
fails.

What we have to do is to make sure that the stream is in the initial state
so we have to call str->reset before we use it for the first time. This
doesn't clash with the moveStart logic because we will get the file position
at the correct place and stream start is aplicate on top of that.

The problem was observed in "primitive print contentstream" test (from
testccontentstream.cc) and it is one of the proper fix for this issue.