xpdf: proper initiazation of XRef fix
Currently we are initializing XRef::streamEnds and XRef::objStr only in the
initInternals method which is called from constructor. This is the proper
place for initialization in general because this code path is generic also
for change revision.

However, when we fail to find PDF header we don't call initInternals and so
those fields end up uninitialized. PDFDoc then crashes in the destructor
because objStr looks like a valid object even though it contains only some
random mess.

One way to solve this would be ignoring the missing PDF header but this may
indicate that the file is not a proper PDF so we could end up with failure
later.

The other way is simply initialize both fields in the constructor header.

This patch fixes one of the issues reported in bt#330.