kernel: fix the outlines tree traversing
Fixes bt#334.

utils::getAllChildrenOfPdfObject which is called from CPdf::getOutlines
relies on existance of the "Next" entry in the outlines tree node
dictionary.

This assumption is not correct because "Next" entry must be present for all
_but_last_ items on the same level. This means that if there is only one
item on the level, there will never by any Next entry.

Document attached to the bt#334 is such an example:
Outlines tree root
	-> 0126 					[124 0]
		-> BACKGROUND 				[125 0]
			-> Instruction			[126 0]
			-> Specific Project Context	[127 0]
			[...]
		-> METHODOLOGY
			-> Identification of Treatment Device Locations
			-> Sizing of Treatment Devices
			[...]
		-> LID DEVICES FOR HOBSONVILLE ROAD WEST
			[...]
		-> CONCLUSIONS AND RECOMMENDATIONS	[137 0]
			[...]


This also removes bogus check for dict == last because we don't have ==
operator for CDict and there is no guarantee that both will be the same
instance. Missing Next entry check is enough here.