Replace explicit Object::free by xpdf::object_deleter
We should not free objects by hand unless it is absolutely necessary and
use a smart pointer with xpdf::object_deleter as destructor instead.

What I consider necessary is free-before-assign-new-value call. As an
example, consider an example:

kernel/ccontentstream.cc: createOperatorFromStream
	[...]
	o->free()
	streamreader.getXpdfObject (*o);
	[...]
free is necessary here, because getXpdfObject fills a new value into the o
and the old value has to be deallocated.

This replacement changes many stack allocated structures to malloc based which
is not good in general but this way we can find those places which require
special treatment more easily and also we have safer handling wrt. memory
leaks.