xpdf: GHash::{add,replace} const variants
GHash supports addition of non-const keys only because it stores this
pointer directly to the GHashBucket. This is, of course, a problem if we
need to store const keys associated with values. Simple approach would be to
change keys to const and change the interface.

The current implementation, however, is even more tricky because the key is
deallocated on cleanup if GHash::deleteKeys is set to true.

The simpliest way how to deal with this is to add overloaded add method
which takes const key and casts it to the simple pointer. To make this
safer we can terminate if deleteKeys is set.

The same applies also to the replace method which silently deallocates the
given key if deleteKeys is set. We just extract functionality into the
helper method with the delete flag.