xpdf: remove unused code from our standard building system
We are not using complete xpdf code as is. E.g. PDFDoc is not used anywhere
in the kernel or GUI and in fact we shouldn't he used at all, because it is
no longer maintained by us (no upstream changes are backported). The same
applies to CoreOutputDev.cc, PDFCore.cc, pdftops.cc, pdftotext.cc,
pdfinfo.cc, pdffonts.cc, pdftoppm.cc, pdfimages.cc.

Let's remove all of them from referenced (in Makefile.in) sources,
distributed headers and includes where-ever they are still used.

We cannot remove all of them, because at least PDFDoc.cc is used by tests,
so we have to create a special target libxpdf-extra which will create
libxpdf-extra.a library and make tests target dependent on it. We can add
also some other files there later.  This library doesn't get distributed.

SecurityHandler includes PDFDoc.h but it uses it only if ENABLE_PLUGINS is
defined. This is not done by our code, but keep it for having clear
compilation in this case.