xpdf: Add EXTRA_XPDF_{C,CXX}FLAGS
We have configurable configure options for CFLAGS and CXXFLAGS for utils,
kernel and gui but none for xpdf code. This shouldn't be any big deal, but
it is nice to have it in sync with others.

Let's turn on -pedantic for xpdf this way to ensure standard enforcement as
this code is also a part of our exported pdfedit-core-dev library.