kernel: getCObjectFromRef fix for trailer's properties
Fixes bt#343

getCObjectFromRef forces that the given property has to have a valid indiref
by means of assert(hasValidRef).

This is, however, problem for all properties which are retrieved from the
trailer as this is a direct object and so it cannot set IProperty::ref to
its children.

Moreover, this function doesn't need this check at all because it doesn't
rely on inidref logic. It just gets the value from the property and
dereference target object from pdf.