Fix libz detection
We have relied on AC_CHECK_FT2 which is setting -lz automaticaly, but this
is not a case for all platforms (e.g. Fedora is different in that regard).

Let's make libz detection explicit. The core autoconf script was taken from
git://git.sv.gnu.org/autoconf-archive.git.

Thanks to Orion Poplawski for reporting.