tools: move tools to src and honor --enable-tools configuration options.
Tools should be placed in src because we want users to play with them (test
directory, on the other hand, is more internal stuff). After we move files
we can honor TOOLS_TARGET and add it to the default targets.

We have to call tools/make clean in all forms of clean target.