xpdf: fixes for nasty const&virtual bugs caused by previous const cleanup
Jozo has reported that some of the functions which were refactored to use
const parameters or which became const broke virtual inheritance because
either its parent or descendant weren't changed so we ended up with an
overloaded virtual function. I tried to check all of those during the
refactoring but some sneaked through, unfortunately. This is really hard to
detect until something goes wrong.

This is rather nasty and leads to subtle bugs when parent method is called
when the descendant is expected depending on the parameters or context.

This patch should fix the remaining places. I have collected all virtual
functions defined in xpdf code and checked each one whether its signature
has changed in the "xpdf: use const for methods and parameters where possible"
commit. If it was changed I have checked all occurrences in the src
directory. I hope this coverage is sufficient and that we have found all of
them.