kernel: prepare OldstylePdfWriter for unexpected generation numbers
OldstylePdfWriter::writeTrailer currently strictly sticks to the
specification and expects XREFROWLENGHT lenght of a xref table entry.
Unfortunately, if some object happen to have generation number longer than 5
digits, we will end up with a invalid xref entry because
snprintf("...%05i n ")

will end up writing as much as possible (which fits into XREFROWLENGHT) from
gen number and `n' will not fit in.

FWIW this shouldn't happen and we try to prevent from that in our code
(objects with gen>MACOBJGEN are not created) but we could get to this from
an already existing object... So we should be rather careful than end up
with an invalid xref table.