xpdf: hack around MAX_INT limit in the XRef::constructXRef
bt#350 contains a document which contains object with object number 225
which has generation number 3452816845 which is out of scope of int type.
PDFSpecification (Appendix C) says that integral numbers have signed 32b
limits and that the generation number is a non-negative number xref table
entry is limited to 5 digits for the generation number so I guess that this
is not allowed.

The document attached to the bug is damaged (xref table entry for object 225
doesn't contain n or f part of the entry) and so XRef::constructXRef is
called. This method tries to reconstruct xref from all objects present in
the document. It uses atoi function for number conversion and so the gen
number for obj 225 is converted to MAX_INT (as it is out of type range).
Therefore XRef::fetch(225, 3452816845) resolves to null and no content is
shown.

Maybe we could ignore this but as other parsing tools of the xpdf code don't
use atoi and parses value whether it overflows or not, maybe we should be
consistent with that and enable these values. This means that we should use
atoll instead.