xpdf: replace atol by C99 compliant strtoll
atol is POSIX and C99 compliant but it doesn't work for 32b machines
as we need because sizeof(long)=sizeof(int)=4B and so we end up with
MAX_INT if the value overflows 31bits.

What we need is to catch the number which overflow signed 32b. We have to use
strtoul instead which, unlike the previous attempt to use atoll, is C99
compliant thus should be OK on Windows as well and we don't end up with
MAX_INT.