kernel: save binary strings as hexstring
Fixes bt#348

Currently we are saving all pdf string objects as literal string. It seems
that this is not 100% correct. In bug#348 we accidentally ended up marking a
stream as changed which caused hexadecimal strings to be converted to
literal strings and Acrobat Reader (and some others) displayed certain
characters incorrectly. The problem seems to be caused by the fact that
PDF specification interprets \r (0x0D or 13) as \n (0x0A or 10) if it is
not preceded by backslash.

Let's make all string which contain non-printable characters as hexadecimal
strings which makes things easier - we don't have to think about back-slashing.

This patch checks the string for any non-printable characters and produce
the hex string for binary strings. This is simple as simpleValueToString is
the only way how to export string objects.