gui: Enable opening read-only documents
Fixed bug 355

If we try to open a read-only document gui will fail because it is opening
each document as ReadWrite or Advanced (depending on the current setting).
This means that CPdf::getInstance opens file with rb+ flags and thus fopen
fails and throws PDFOpenException.
We can prevent from that simply by falling back to ReadOnly if ReadWrite or
Advanced modes fail.