configure: check for libpng and its header if tools are configured
This fixes bt#357

add_image tool uses png and its header but none of them are checked during
configure. Let's add a dependency on both (the library and the header) if
tools are configured.

Do not care much about parameters for a special path configuration or
special c flags for now. We can do it as per request.