kernel: fix bad TextSimpleOperator::getRawText
getRawText is intended only for low level operations when one wants to work
directly with the font dependent data (character codes) which can be different
than what is printed in the end (typical examples are pdfs created by
OpenOffice which uses char codes from 1 to the number of distinct characters
and each character is then assigned to one such a code in the font unicode map).

CContentStream::replaceText is not such a function, obviously, so it has to
use getFontText instead.