kernel: fix TextSimpleOperator::setRawText for TJ operator
OMG we have this "am to lazy to do it" for 4 years in the code... :/

Anyway, let's add an implementation for it now, finally. TJ operators eats
an array of strings which can be mixed with numbers used for formatting.
This method is highlevel in the regards of formatting so we can safely get
rid of all (but one) elements of the array and use just the given string as
the only array element.