xpdf: implement font based Unicode to Char code translation
PDF text operators expect char code strings which are translated to the
unicode characters according to the used fonts when the string is displayed
(by means of GfxFont::getNextChar method). GfxFont doesn't implement this
method directly and it relies on descendants to do that.

Both Gfx8BitFont and GfxCIDFont use CharCodeToUnicode class for translation
(mapToUnicode method) to get unicode from char code.  Let's add an inverse
method mapFromUnicode to the CharCodeToUnicode which will do the reverse
mapping (from unicode to char codes) and use it in GfxFont directly. We rely
on the proper unicode and its size there.

WARNING:
I have tested this only for Gfx8BitFont only and even for that one only for
1B unicodes only which use map rather than sMap.