kernel: replace TextSimpleOperator::setRawText by setFontText
There is no real low level user of setRawText because the only caller
CContentStream::replaceText is intended to replace a text by another text
and doesn't care about fonts and encoding. So in fact it wants to use
something like setFontText.

Let's update setRawText to do the String -> font specific char codes
transformation and rename it to setFontText.