kernel: enable CContentStream withtout streams
Currently we need streams in order to create a content stream instance.
This makes sense but there are cases when we need an instance and we don't
have streams ready yet (e.g. when we are creating operands for a text
operator and we need content stream to provided Gfx* internals).

Let's do this change separately from the rest for easier reviews. The stream
can be later supplied by setStreams method which is responsible for a proper
cleanup from the previous state as well.