kernel: refactor CContentStream instantiation
CContentStream instantiation involves some non-trivial steps (like graphical
state and resources objects creation and setting the self reference). Let's
wrap this into a static helper function (this has to be a class function
because page::display - used internally - is not accessible normally) and
replace all direct instantiation with it.

Finally, addToFront resp. addToBack has a new variant with the content stream
parameter. This is intended for later use but it is nice to have this change
along with the instantiation changes.