kernel: add createOperatorText helper
Our standard general purpose helper createOperator is not sufficient for
simple text operators creation. The main obstacle is that it doesn't have
sufficient information about font which is necessary for proper
unicode->char code conversion. This means that createOperator for text
operators is usable only for a low level usage where the operands already
contain mapped values.

Let's add createOperatorText helper which contains the complete information
to do the proper mapping and which then simply calls the low-level
createOperator routine. In order to do that we need the content stream which
contains GfxResources which contains all fonts and fontName. text is
expected to be unicode encoded.