kernel: add {CArray,CDict}::getproperty<Type> method
Currently both CArray and CDict provide only
shared_ptr<IProperty> getProperty(id) method which is OK but it is far from
being convenient for usage. You always have to cast and declare more
variables than necessary:

shared_ptr<IProperty> p = dict->getProperty(name);
shared_ptr<Type> pType = IProperty::getSmartCObjectPtr<Type>(p);

This gets even more complicated if the property is a reference and we need
to access the target object. Sure we have a helper function for that but
there are too many of them and it is not trivial to find the one you want to
use (grep simple getTypeFrom{Dictionary,Array} shows that even we are not
using them much).

Let's add a template getProperty<Type> method which does the thing and it is
directly in the C{Dict,Array} so they are easy to find.