kernel: makeStringPdfValid do not escape already escaped string
Fixes bt#359

TestCContentStream fails with:
1) test: (anonymous namespace)::TestCContentStream::TestSetCS (F) line: 93 testccontentstream.cc
assertion failed
- Expression: tmp1 == tmp2

The problem is in makeStringPdfValid function which escapes the given string
even though it is already escaped.

The easiest way to fix this is to process the given string from the end and
escape only if the previous (the next processed) character is not `\'.

This patch also removes iterator parameters and it replaces it by
std::string reference because iterators are not practical for reverse
processing and reverse iterator wouldn't be very nice.