use unsigned char for isCTYPE function family
isprint and others from ctype function family take the signed integer as the
argument but we are providing the signed char in there at many places. This may
result in an overflow for characters which don't fit into 7b and become a
negative number. glibc does some trickery to overcome this problem but others
can have problems.

According to a report by Ravikant Dubey, he is seeing runtime assert
failures when he is running the code under Microsoft Visual C++ 2008 with
debugging turned on:

Expression: (unsigned)(c+1) <= 256


Thanks go to Jozo who come up with the runtime assert suspicion.