build: do not use -I. in the include path
Using -I. is nasty and can lead to a silent breakage when we include
something different than we want.

Moreover we do not need this hack anymore because all includes have been
converted to be source tree root based for the pdfedit-core-dev.

Unfortunately, we cannot do anything about QT based automatically
generated Makefiles which contain that by default.