configure: make portability and c/c++ standard compilation flags configurable
We had a report that compilation fails on some versions of Cygwin. It turned
out that the problem is in -std=c++98 parameter forces a behavior (probably
a bug) which leads that some posix functions are not defined (namely mkstemp,
fdopen and probably others).

The easiest way to fix this is to provide a way to disable those parameters.
This should not have any effect on the result of the compilation.

This patch also adds -D_XOPEN_SOURCE=500 as both function require it.