makedist script refactoring
Let's simplify makedist script and make it git aware. Previous workflow
was a bit awkward. You would need to run /tools/make_release to get CVS
exported tree (without CVS internals) which called makedist in a
separate directory.

First of all we do not need to export the tree because git ls-tree along
with tar and little bit of magic can do the same thing directly from the
tree.  We can drop release_stamp and replace it by .scmversion because
the version is sufficient (./getversion -v -r is unambiguous).

We can also drop make_release hack which was just a CVS hack around.